# [http/spotify] Allow reuse of curl handle to improve Spotify scan performance