My Spotify library is getting larger and it takes now a significant amount of time to run the spotify library scan ...

One way I found to improve it is to reuse the curl handle. Reusing the handle - instead of creating a new one for each request - will keep the connections, session ID cache, DNS cache and cookies.

The implementation in this PR is a proof of concept.

Some "benchmarks" from my raspberry pi 3B:

Without reusing the curl handle (current master branch):

```
[2022-04-17 10:19:28] [  LOG]  spotify: Spotify scan completed in 313 sec
[2022-04-17 10:52:57] [  LOG]  spotify: Spotify scan completed in 212 sec
[2022-04-17 10:57:36] [  LOG]  spotify: Spotify scan completed in 248 sec
```

With reusing the curl handle (this PR branch):

```
[2022-04-17 11:10:09] [  LOG]  spotify: Spotify scan completed in 145 sec
[2022-04-17 11:13:45] [  LOG]  spotify: Spotify scan completed in 200 sec
[2022-04-17 11:17:12] [  LOG]  spotify: Spotify scan completed in 194 sec
```

(On my laptop the change has no measurable performance impact)

I am not sure how to best implement it. I was thinking instead of passing an optional curl handle to `http_client_request`, maybe it would be better to keep a curl handle per thread and make the use of it transparent to the caller of `http_client_request`. 

@ejurgensen I would appreciate your thoughts on this.