@ejurgensen 2022-04-25
Sounds good with improved scan times, but I think http.c should fully abstract curl, so I suggest a solution where there is no direct reference to curl in spotify_webapi.

I guess the next step in improving performance would be to make the calls parallel, i.e. make them async, and I imagine per thread handles won't be good for that. I guess the standard C approach is also to have the caller hold all context, if possible.