@chme 2022-04-26
@ejurgensen Could you take another look at this? I am not sure if this is what you had in mind, I am open to suggestions on how to improve it. And I agree that keeping a curl handle per thread is not a good way to do it.

I don't think, we should do parallel requests. I imagine this will open up other issues and introduce lots of complexity (e. g. DB lockups). What I think could have a big performance improvement is to better identify if data has changed. I am currently experimenting with the "If-None-Match" header to get a 304 response from Spotify, if nothing has changed since the last scan. I hope to have some experimental code finished soonish.