@ejurgensen 2022-04-29
Looks good to me. One minor suggestion is to organize spotifywebapi_init() the "standard" way. So instead of:

```
 ret = spotify_init();

 http_client_session_init(&session);
 return ret;
```

then

```
 ret = spotify_init();
 if (ret < 0)
   return -1; (or return ret)

 http_client_session_init(&session);

 return 0;
```